Refactoring of emmy clients
This commit refactors client-side of emmy as it introduces protocol-specific
clients and removes obsolete structs like Client and ProtocolParams. It also
removes centralized instantiation of protocol-specific clients from client.go
and moves this logic to emmy.go instead. For functionality concerning the
communication aspect, which is shared across several protocol clients, we are
now embedding a "common" struct GenericClient. This fixes xlab-si/emmy#3

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>