Improve emmy CLI
This commit introduces an improved CLI. All the CLI-related
functionality is moved to the 'cli' package. Main function in
emmy.go is greatly simplified. The example command was removed
from CLI due to its complexity and questionable usability.
Protocols are now run as client subcommands. Finally, we
introduce protocol-level flags that were fixed until now,
as well as confiruration of server's port and endpoint via
CLI.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>