Don't panic when server port is occupied
This commit introduces a simple fix that reports an error
instead of panicking when the server cannot be started due
to an error emitted by net.Listen(port) call. This happens,
for instance, when a port we want our server to listen on
is occupied. The server CLI command now catches such errors
and displays them to the user.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>