Remove leftover protobuf files and definitions
This commit modifies the files in the protobuf package. A file
left over from a previous merge was removed. Definitions of services
in the msgs.proto were cleaned up so that only the set we are
actually using is defined (e.g. the leftover test service and RPC
definitions were cleaned up). A new msgs.pb.go was generated based
on this cleaned up file.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>