Expose functions for generation of master secret keys
Until now client's master secret key was loaded from the config
file. This commit adds two simple functions for secret key
generation, one hooked to PseudonymsysClient, the other to
PseudonymsysClientEC, which the client can call to generate the
secret key on-the-fly. Fixed key and corresponding function for
loading it were removed from the config package, and the tests
were updated accordingly.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>