Move crypto tests to crypto packages and refactor them
This is the first commit in the series of commits aiming towards
retirement of the common test package. It moves existing crypto
tests from various files within our test package to appropriate
packages within crypto directory. Tests were also refactored in
order to make use of testing package's reporting capabilities
instead of relying on fmt and log packages. Travis CI file was
also updated so that all the tests in the repository are run
instead of just those in the test package.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>