Make client connections more configurable
This commit introduces a ConnectionConfig struct that holds configuration
information required for client to contact the server: endpoint,
CA certificate and TLS insecurity parameter are now passed to client's
GetConnection function within a single struct. Moreover, a large part
of this commit is allowing for more flexibility with respect to CA
certificate, which is now passed as a byte array instead of
having to pass path to certificate on the client's filesystem. Some
additional tests were also added in client/tls_test.go.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>