Add redis address CLI flag
This commit adds a new "db" flag to emmy server CLI command. The
flag should point to the address of the redis database that will
hold registration keys. This allows us to override the default
value, read from defaults.yml, at server startup. The address is
now also passed to the initializer of the Server struct. Moreover,
this commit also increases severity of not being able to connect
to redis at a given address to "critical", and improves the error
message that gets reported.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>