Retire insecure connections in favor of serverNameOverride
This commit removes Insecure field from ConnectionConfig struct and
replaces it with ServerNameOverride string field. When Insecure was
set to true, server's certificate validation was completely bypassed,
even though only hostname verification was problematic. This commit
retires Insecure field as it is a security risk. Instead, it allows
specifying a serverNameOverride that allows us to establish a secure
TLS connection even when the server's hostname does not match server
certificate's CN (it is now required that serverNameOverride, not
server hostname, matches server certificate's CN).

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>