Support for less generic services and RPCs (#54)
The purpose of this commit is to allow for a better definition of the
protobuf services and their RPCs. Up until now, we had all the
cryptographic schemes and protocols implemented with a single
service and RPC. We now have separate RPCs for several
protocols within the pseudonym system and pseudonym system
CA schemes. Server-side handlers were updates, client-side
implementations now use reflection for opening client streams.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>