Make client connections more flexible (#55)
This commit introduces a ConnectionConfig struct that holds configuration
information required for client to contact the server: endpoint,
CA certificate and ServerNameOverride parameter are now passed to
client's GetConnection function within a single struct. CA certificate is 
now passed as a byte array. The Insecure connetion option was retired
in favor of ServerNameOverride that allows us to establish a secure
TLS connection even when the server's hostname does not match server
certificate's CN. CLI, tests and README were updated.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>