Add Dockerfile and docker-compose file (#61)
This commit adds a Dockerfile for building emmy server image. In
addition, docker-compose file is added that starts emmy server
container as well as redis container. These are added solely for
convenience during development; their use is not assumed during
unit testing on a given host or for for Travis CI build. Docs for
development were also updated with brief instructions.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>