Modify configuration of default schnorr groups (#63)
This commit removes loading default configuration where it is not needed. 
Shcnorr groups in test codes are now generated on-the-fly, since we don't
share the parameters between entities (clients and server) for extended
periods of time. In addition, clients and server now use a simplified version
of the function for loading schnorr group, config.LoadSchnorrGroup().