Add keypair generation functions for pseudonym system
This commit adds functions for generation of keypairs to be used
with pseudonym system scheme (both EC and modular). Currently these
functions are unused, as server-side relies on using fixed values
that are loaded from config, so using these functions in test code
breaks the tests. When configuration will be improved, these functions
will be called for initial setup of pseudonym system CA and
organizations.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>