New structs for pseudonym system keys (#75)
This commit adds definitions of PubKey, PubKeyEC and SecKey, which
correspond to public keys for both implementations of pseudonym
system scheme, and their secret keys, which are of the same type
for both implementations. It replaces old OrgPubKeysEC and OrgPubKeys
structs with newly introduced ones. Moreover, functions and structs that did
not accept key parameters as structs were updated. Changes are propagated
to all the affected packages.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>