Use dep for dependency management (#74)
This commit introduces the use of golang's tool for dependency
management called dep to enable strict version control for emmy's
dependencies. Dep uses the Gopkg.toml file to manage dependencies, which
are stored in the vendor directory. Some changes to Makefile, Dockerfile and
.travis.yml were needed to make everything work with the vendor directory.
The changes to the workflow when developing Emmy are described in 
docs/develop.md.

Signed-off-by: Jan Hartman <jan.hartman@xlab.si>