Remove ProtocolError message
This commit removes custom ProtocolError protobuffer message that
was a part of our Message message type. Instead, when a protocol
error happens, we simply return the status.Error from the RPC, which
is a standard way of reporting RPC errors to clients. As a consequence,
client's send and server's recieve functions are simplified. The change
also affects pseudonym system communication logic, which is now also
simplified.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>